add stub for trace stats

+++ trace-utils/app/tencent/split.cpp
@@ -7,8 +7,6 @@
 #include <tuple>
 #include <chrono>
 
-#include <omp.h>
-
 #include <archive.h>
 #include <archive_entry.h>
 
@@ -45,8 +43,10 @@ SplitApp::SplitApp(): App(split::name, split::description) {
 }
 
 SplitApp::~SplitApp() {
-    log()->info("Removing temporary directory", tmp_dir_path);
-    fs::remove_all(tmp_dir_path);
+    if (!tmp_dir_path.empty()) {
+        log()->info("Removing temporary directory", tmp_dir_path);
+        fs::remove_all(tmp_dir_path);
+    }
     indicators::show_console_cursor(true);
 }
